test: move basic.rs tests into sync.rs, cover PNG MIME in unit test
`initial_sync` duplicated `basic_deploy` and used a 60-line hand-rolled
PNG generator solely to check content-type detection, which is purely
extension-based and belongs in a unit test. Replace it with a
`load_infers_png_mime` unit test in `assets-sync/src/content.rs`, delete
`generate_minimal_png`, consolidate `basic.rs` into `sync.rs`, and
update TEST.md accordingly.
