+++ e2e/src/lib.rs
@@ -129,71 +129,3 @@ pub fn list_assets(project: &Path) -> Vec<AssetDetails> {
         .expect("failed to decode candid response");
     assets
 }
-
-/// Generate a minimal 1×1 grayscale PNG at runtime without any external crate.
-///
-/// Uses a stored (non-compressed) deflate block inside zlib so no compression
-/// library is needed.  The CRC32 and Adler-32 checksums are computed inline.
-pub fn generate_minimal_png() -> Vec<u8> {
-    fn crc32(data: &[u8]) -> u32 {
-        let mut crc = !0u32;
-        for &b in data {
-            let mut v = (crc ^ b as u32) & 0xFF;
-            for _ in 0..8 {
-                v = if v & 1 != 0 {
-                    0xEDB88320 ^ (v >> 1)
-                } else {
-                    v >> 1
-                };
-            }
-            crc = (crc >> 8) ^ v;
-        }
-        !crc
-    }
-
-    fn adler32(data: &[u8]) -> u32 {
-        let (mut s1, mut s2) = (1u32, 0u32);
-        for &b in data {
-            s1 = (s1 + b as u32) % 65521;
-            s2 = (s2 + s1) % 65521;
-        }
-        (s2 << 16) | s1
-    }
-
-    fn chunk(tag: &[u8; 4], data: &[u8]) -> Vec<u8> {
-        let crc = crc32(&[tag.as_slice(), data].concat());
-        let mut v = Vec::new();
-        v.extend_from_slice(&(data.len() as u32).to_be_bytes());
-        v.extend_from_slice(tag);
-        v.extend_from_slice(data);
-        v.extend_from_slice(&crc.to_be_bytes());
-        v
-    }
-
-    // 1×1 grayscale scanline: filter byte 0, pixel value 128
-    let scanline: &[u8] = &[0x00, 0x80];
-
-    // Wrap the scanline in a zlib stored (non-compressed) deflate block.
-    let len = scanline.len() as u16;
-    let mut zlib = Vec::new();
-    zlib.extend_from_slice(&[0x78, 0x01]); // zlib header (deflate, default compression check)
-    zlib.push(0x01); // BFINAL=1, BTYPE=00 (stored block)
-    zlib.extend_from_slice(&len.to_le_bytes()); // LEN
-    zlib.extend_from_slice(&(!len).to_le_bytes()); // NLEN (one's complement)
-    zlib.extend_from_slice(scanline);
-    zlib.extend_from_slice(&adler32(scanline).to_be_bytes()); // zlib checksum
-
-    let mut png = Vec::new();
-    png.extend_from_slice(&[0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A]); // PNG signature
-
-    // IHDR: 1×1, 8-bit grayscale (color_type=0)
-    let mut ihdr = Vec::new();
-    ihdr.extend_from_slice(&1u32.to_be_bytes()); // width
-    ihdr.extend_from_slice(&1u32.to_be_bytes()); // height
-    ihdr.extend_from_slice(&[8, 0, 0, 0, 0]); // bit_depth=8, color_type=0, compress=0, filter=0, interlace=0
-    png.extend_from_slice(&chunk(b"IHDR", &ihdr));
-
-    png.extend_from_slice(&chunk(b"IDAT", &zlib));
-    png.extend_from_slice(&chunk(b"IEND", &[]));
-    png
-}
